+++ CHANGELOG.md
@@ -2,11 +2,12 @@
 
 ## V4.2.6
 
-#### Date: 6/19/2026
+#### Date: 6/22/2026
 
 ### :bug: Bug Fixes
 
 -   Fixed an issue where Anthropic models might under-report credit usage when streaming chat messages.
+-   Fixed an issue where some HTTP request paths would cause the server to crash.
 
 ## V4.2.5