+++ CHANGELOG.md
@@ -9,6 +9,13 @@
 -   Fixed an issue where Anthropic models might under-report credit usage when streaming chat messages.
 -   Fixed an issue where some HTTP request paths would cause the server to crash.
 
+### :rocket: Features
+
+-   Added new BIOS option.
+    -   `public` - Works exactly like `public inst`.
+-   BIOS option label updated to `public` (for example: `public 24h` or `public 45m`)
+
+
 ## V4.2.5
 
 #### Date: 6/10/2026